+++ protected/humhub/libs/SelfTest.php
@@ -285,6 +285,39 @@ class SelfTest
             ];
         }
 
+        // Checks ImageMagick Extension
+        $title = 'PHP - ImageMagick Extension';
+
+        if (class_exists('Imagick', false)) {
+            $checks[] = [
+                'title' => Yii::t('base', $title),
+                'state' => 'OK'
+            ];
+        } else {
+            $checks[] = [
+                'title' => Yii::t('base', $title),
+                'state' => 'WARNING',
+                'hint' => 'Optional'
+            ];
+        }
+
+        // Checks GraphicsMagick Extension
+        $title = 'PHP - GraphicsMagick Extension';
+
+        if (class_exists('Gmagick', false)) {
+            $checks[] = [
+                'title' => Yii::t('base', $title),
+                'state' => 'OK'
+            ];
+        } else {
+            $checks[] = [
+                'title' => Yii::t('base', $title),
+                'state' => 'WARNING',
+                'hint' => 'Optional'
+            ];
+        }
+
+
         // Checks LDAP Extension
         $title = 'LDAP Support';